refactor: Role(Enum) 역할에 맞게 global -> domain 이동 #8
