+++ src/main/java/KUSITMS/WITHUS/domain/user/dto/UserRequestDTO.java
@@ -1,6 +1,6 @@
 package KUSITMS.WITHUS.domain.user.dto;
 
-import KUSITMS.WITHUS.global.auth.enumerate.Role;
+import KUSITMS.WITHUS.domain.user.enumerate.Role;
 import io.swagger.v3.oas.annotations.media.Schema;
 
 @Schema(description = "회원 관련 요청 DTO")
