+++ src/main/java/KUSITMS/WITHUS/domain/user/enumerate/Role.java
@@ -1,4 +1,4 @@
-package KUSITMS.WITHUS.global.auth.enumerate;
+package KUSITMS.WITHUS.domain.user.enumerate;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
