+++ src/main/java/KUSITMS/WITHUS/global/auth/jwt/JwtFilter.java
@@ -1,7 +1,7 @@
 package KUSITMS.WITHUS.global.auth.jwt;
 
 import KUSITMS.WITHUS.global.auth.dto.CustomUserDetails;
-import KUSITMS.WITHUS.global.auth.enumerate.Role;
+import KUSITMS.WITHUS.domain.user.enumerate.Role;
 import KUSITMS.WITHUS.domain.user.entity.User;
 import KUSITMS.WITHUS.global.exception.CustomException;
 import KUSITMS.WITHUS.global.exception.ErrorCode;
